This is the title of the git commit
This is the message part. I relay my future intentions here, ie what I
want to see happen further. as signposts for future work. So This was a
commit to change the event to events file. I need the Event file to
create a view for editing event. This will be for editing the title,
indicators and indicator strength, tags. Each time an indicator is
updated it is recorded in the log, so one can later interrogate the
changes in indicator strength over time.