Adding the loading spinner with message
This spinner is shown when the data is first loaded from the heroku
database via graphql. Once Apollo client has a cache it should not show
his message again. There is I suppose an edge case whereby the database
might never respond, the internet is down, the database is down and god
knows a whole lot of other stuff. I thin the easiest fix is to time out
after say around 15 seconds and pass a message to use to say this seems
to be taking longer than usual, then decide where to go from there.