Add SourceAdminForm
So I added the source admin form. At this stage it is only a bare bones
form using HTML5 controls. There still needs Formik to be added with its
controlled component controls. And at this stage it will be only using a
basic comma seperated list for the keywords. Going forward I intend to
refactor this and to allow for a more fancy typeahead component which
automatically adds keywords to a list and then uses this list. The
lookup for the typeahead can come from previous keywords used. I.e.
every keyword used in the past gets saved into a default_keywords_list.
The actual form also needs to be in a modal, much like all the other
forms will be in the system. I intend to create a app that follows
Notions nice clean design.